requirements.txt
# Core
fastapi
uvicorn

python-dotenv

# Pydantic
pydantic-settings
pydantic  # Required for extra types

# Environment
python-dotenv  # Optional but recommended

# Google Calendar API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# LLM Provider (e.g. Gemini)
google-generativeai  # Only if using Gemini

startette
